[ksmart46team04] author: 김가람 file: ks46team04/src/main/java/ks46team04/common/controller/CommonController.java ks46team04/src/main/resources/templates/fragments/header2.html ks46team04/src/main/resources/templates/admin/user/login.html ks46team04/src/main/resources/templates/admin/user/removeUser.html ks46team04/src/main/resources/static/css/sb-admin-2.min.css ks46team04/src/main/java/ks46team04/admin/controller/UserController.java ks46team04/src/main/java/ks46team04/admin/service/UserService.java

+++ ks46team04/src/main/java/ks46team04/admin/service/UserService.java
@@ -57,17 +57,16 @@ public class UserService {
 	}
 	
 	
-	public void removeUser(String userId) {
+
+	public void removeUser(String userId, String userPw) {
 		User userInfo = userMapper.getUserInfoById(userId);
 		if(userInfo != null) {
-			String userLevel = userInfo.getUserLevel();
-			switch (userLevel) {
-			case "2":
-				
-				// 회원 탈퇴
-				userMapper.removeUserById(userId);		
+			String checkPw = userInfo.getUserPw();
+			if (checkPw.equals(userPw)) {
+				userMapper.removeUserById(userId);
 			}
-		}	
+		}
+		
 	}
 	
 	public void modifyUser(User user) {
